test and lint examples
