+++ examples/simple/src/mock_backend/backend_state.rs
@@ -16,7 +16,7 @@ pub fn new_ed25519_verifier(base_url: &str) -> BackendState<Ed25519Verifier> {
             MemoryCache::default(),
             ReqwestFetcher::default().with_base_url(base_url),
         ),
-        Ed25519Verifier::default(),
+        Ed25519Verifier {},
     )
 }
